backend/parser/requirements.txt
pandas
pymongo
python-dotenv
xlsx2csv 